backend/requirements.txt
Django==2.2.16
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
django-filter==22.1
drf-extra-fields==3.4.0
gunicorn==20.0.4
PyJWT==2.1.0
Pillow==9.2.0
psycopg2-binary==2.8.6